Create a long data.frame from a MultiAssayExperiment counts object

This function takes a MultiAssayExperiment object and a specified taxon level of interest and creates a long data.frame that can be used more easily for plotting counts data.

Usage

get_long_data(dat, taxon_level, log = FALSE, counts_to_CPM = FALSE)

Arguments

dat

A MultiAssayExperiment object specially formatted as an animalcules output.

taxon_level

Character string indicating the level of taxonomy to aggregate the counts data. Must be the name of a column in MultiAssayExperiment::rowData(dat).

log

logical. Indicate whether an assay returned should be the log of whichever assay is specified in "output_name". If counts_to_CPM = TRUE as well, then a log CPM assay will also be created. Default is FALSE.

counts_to_CPM

logical. This argument only applies if the input_type is a counts assay. If TRUE, then the output assays will include a normalized CPM assay. If log = TRUE as well, then a log CPM assay will also be created. Default is TRUE.

Value

A data.frame consisting of the counts data, taxa, and metadata.

Examples

in_dat <- system.file("extdata/MAE_small.RDS", package = "LegATo") |> readRDS()
out <- get_long_data(in_dat, "genus", log = TRUE, counts_to_CPM = TRUE)
#> Registered S3 method overwritten by 'rmutil':
#>   method         from
#>   print.response httr
head(out)
#>           taxon Sample Subject    Sex Month Group Pairing HairLength      Age
#> 1 Acinetobacter    X-1      S1 Female     1     A       1   54.59018 37.60217
#> 2 Acinetobacter    X-2      S2 Female     1     B       1   69.32642 66.77243
#> 3 Acinetobacter    X-3      S3 Female     1     A       2   42.91915 53.23820
#> 4 Acinetobacter    X-4      S4 Female     1     B       2   53.18740 52.56244
#> 5 Acinetobacter    X-5      S5 Female     1     A       3   44.28454 60.37946
#> 6 Acinetobacter    X-6      S6 Female     1     B       3   37.66513 42.47814
#>   Sample_id Abundance
#> 1       X-1      4484
#> 2       X-2     19992
#> 3       X-3    150783
#> 4       X-4       886
#> 5       X-5      1008
#> 6       X-6      1928