Create a long data.frame with grouped abundances from a MultiAssayExperiment counts object

This function takes a MultiAssayExperiment object and a specified taxon level of interest and creates a long data.frame that can be used more easily for plotting counts data in a stacked bar plot or a stacked area chart. The function groups taxa and computes relative abundance within taxa strata.

Usage

get_stacked_data(dat, taxon_level = "genus", covariate_1, covariate_time)

Arguments

dat

A MultiAssayExperiment object specially formatted as an animalcules output.

taxon_level

Character string indicating the level of taxonomy to aggregate the counts data. Must be the name of a column in MultiAssayExperiment::rowData(dat).

covariate_1

Character string, the name of the covariate in `dat` by which to color and group samples. Default is NULL.

covariate_time

Character string giving the name of the discrete time-based covariate in the metadata to group abundances by.

Value

A data.frame consisting of the counts data, taxa, and metadata.

Examples

in_dat <- system.file("extdata/MAE_small.RDS", package = "LegATo") |> readRDS()
get_stacked_data(in_dat, "genus", covariate_1 = "Sex", covariate_time = "Month")
#>                  taxon covariate_1 covariate_t        value
#> 1        Acinetobacter      Female           1 5.549263e-01
#> 2            Aeromonas      Female           1 1.579758e-02
#> 3     Alkalihalophilus      Female           1 1.042267e-03
#> 4             Brucella      Female           1 1.057197e-02
#> 5      Corynebacterium      Female           1 5.915736e-02
#> 6        Cutibacterium      Female           1 2.979282e-03
#> 7              Delftia      Female           1 2.440755e-06
#> 8         Enterobacter      Female           1 5.800568e-04
#> 9          Escherichia      Female           1 3.678564e-02
#> 10         Haemophilus      Female           1 6.209398e-04
#> 11             Kocuria      Female           1 6.662898e-04
#> 12            Listeria      Female           1 2.055714e-02
#> 13           Moraxella      Female           1 6.377616e-02
#> 14          Mycoplasma      Female           1 1.965690e-02
#> 15  Noviherbaspirillum      Female           1 1.192899e-04
#> 16               Other      Female           1 8.867107e-04
#> 17          Paracoccus      Female           1 5.876816e-02
#> 18         Pseudomonas      Female           1 3.701786e-02
#> 19      Staphylococcus      Female           1 1.020357e-01
#> 20       Streptococcus      Female           1 8.279496e-03
#> 21           Weissella      Female           1 5.772429e-03
#> 22       Acinetobacter        Male           1 7.653653e-01
#> 23           Aeromonas        Male           1 6.454990e-03
#> 24    Alkalihalophilus        Male           1 4.187020e-04
#> 25            Brucella        Male           1 6.202516e-03
#> 26     Corynebacterium        Male           1 8.361092e-02
#> 27       Cutibacterium        Male           1 7.269789e-03
#> 28             Delftia        Male           1 0.000000e+00
#> 29        Enterobacter        Male           1 6.140963e-03
#> 30         Escherichia        Male           1 1.465457e-03
#> 31         Haemophilus        Male           1 1.453136e-03
#> 32             Kocuria        Male           1 0.000000e+00
#> 33            Listeria        Male           1 2.083036e-02
#> 34           Moraxella        Male           1 1.994987e-03
#> 35          Mycoplasma        Male           1 7.195907e-03
#> 36  Noviherbaspirillum        Male           1 0.000000e+00
#> 37               Other        Male           1 0.000000e+00
#> 38          Paracoccus        Male           1 1.985546e-02
#> 39         Pseudomonas        Male           1 5.732503e-03
#> 40      Staphylococcus        Male           1 5.666421e-02
#> 41       Streptococcus        Male           1 6.327726e-03
#> 42           Weissella        Male           1 3.017109e-03
#> 43       Acinetobacter      Female           2 2.437613e-01
#> 44           Aeromonas      Female           2 1.391922e-02
#> 45    Alkalihalophilus      Female           2 4.326438e-04
#> 46            Brucella      Female           2 2.843060e-02
#> 47     Corynebacterium      Female           2 2.906003e-01
#> 48       Cutibacterium      Female           2 4.794796e-03
#> 49             Delftia      Female           2 6.781684e-05
#> 50        Enterobacter      Female           2 5.535890e-04
#> 51         Escherichia      Female           2 9.204748e-03
#> 52         Haemophilus      Female           2 1.273125e-03
#> 53             Kocuria      Female           2 4.993988e-04
#> 54            Listeria      Female           2 7.013477e-02
#> 55           Moraxella      Female           2 1.312362e-02
#> 56          Mycoplasma      Female           2 6.402683e-03
#> 57  Noviherbaspirillum      Female           2 0.000000e+00
#> 58               Other      Female           2 7.413018e-04
#> 59          Paracoccus      Female           2 7.040139e-02
#> 60         Pseudomonas      Female           2 2.319198e-02
#> 61      Staphylococcus      Female           2 1.932255e-01
#> 62       Streptococcus      Female           2 2.539212e-02
#> 63           Weissella      Female           2 3.849096e-03
#> 64       Acinetobacter        Male           2 2.338258e-01
#> 65           Aeromonas        Male           2 2.973978e-03
#> 66    Alkalihalophilus        Male           2 0.000000e+00
#> 67            Brucella        Male           2 5.086937e-02
#> 68     Corynebacterium        Male           2 3.780536e-01
#> 69       Cutibacterium        Male           2 6.791380e-03
#> 70             Delftia        Male           2 4.247027e-04
#> 71        Enterobacter        Male           2 3.747377e-04
#> 72         Escherichia        Male           2 0.000000e+00
#> 73         Haemophilus        Male           2 9.993005e-05
#> 74             Kocuria        Male           2 0.000000e+00
#> 75            Listeria        Male           2 7.787796e-02
#> 76           Moraxella        Male           2 1.226766e-02
#> 77          Mycoplasma        Male           2 1.739804e-03
#> 78  Noviherbaspirillum        Male           2 0.000000e+00
#> 79               Other        Male           2 0.000000e+00
#> 80          Paracoccus        Male           2 1.469100e-01
#> 81         Pseudomonas        Male           2 5.504251e-03
#> 82      Staphylococcus        Male           2 4.970205e-02
#> 83       Streptococcus        Male           2 2.749029e-02
#> 84           Weissella        Male           2 5.094501e-03
#> 85       Acinetobacter      Female           3 4.473384e-01
#> 86           Aeromonas      Female           3 4.808669e-03
#> 87    Alkalihalophilus      Female           3 1.118665e-04
#> 88            Brucella      Female           3 2.775884e-02
#> 89     Corynebacterium      Female           3 2.163773e-01
#> 90       Cutibacterium      Female           3 1.254615e-03
#> 91             Delftia      Female           3 1.491449e-04
#> 92        Enterobacter      Female           3 3.696449e-04
#> 93         Escherichia      Female           3 2.238329e-04
#> 94         Haemophilus      Female           3 2.085964e-06
#> 95             Kocuria      Female           3 8.285828e-05
#> 96            Listeria      Female           3 6.919301e-02
#> 97           Moraxella      Female           3 4.387693e-03
#> 98          Mycoplasma      Female           3 1.039901e-02
#> 99  Noviherbaspirillum      Female           3 2.213423e-05
#> 100              Other      Female           3 0.000000e+00
#> 101         Paracoccus      Female           3 9.260826e-02
#> 102        Pseudomonas      Female           3 7.211173e-03
#> 103     Staphylococcus      Female           3 9.381225e-02
#> 104      Streptococcus      Female           3 1.906107e-02
#> 105          Weissella      Female           3 4.828210e-03
#> 106      Acinetobacter        Male           3 9.373308e-01
#> 107          Aeromonas        Male           3 2.760035e-04
#> 108   Alkalihalophilus        Male           3 0.000000e+00
#> 109           Brucella        Male           3 7.346110e-04
#> 110    Corynebacterium        Male           3 8.611660e-03
#> 111      Cutibacterium        Male           3 1.326811e-05
#> 112            Delftia        Male           3 0.000000e+00
#> 113       Enterobacter        Male           3 0.000000e+00
#> 114        Escherichia        Male           3 2.760035e-04
#> 115        Haemophilus        Male           3 0.000000e+00
#> 116            Kocuria        Male           3 8.845409e-06
#> 117           Listeria        Male           3 1.563372e-03
#> 118          Moraxella        Male           3 2.624004e-02
#> 119         Mycoplasma        Male           3 6.059130e-03
#> 120 Noviherbaspirillum        Male           3 0.000000e+00
#> 121              Other        Male           3 0.000000e+00
#> 122         Paracoccus        Male           3 1.483868e-02
#> 123        Pseudomonas        Male           3 5.520069e-04
#> 124     Staphylococcus        Male           3 2.192964e-03
#> 125      Streptococcus        Male           3 4.654292e-04
#> 126          Weissella        Male           3 8.372311e-04
#> 127      Acinetobacter      Female           4 2.202415e-01
#> 128          Aeromonas      Female           4 5.550086e-03
#> 129   Alkalihalophilus      Female           4 5.488474e-04
#> 130           Brucella      Female           4 3.940695e-02
#> 131    Corynebacterium      Female           4 4.383634e-01
#> 132      Cutibacterium      Female           4 1.585227e-04
#> 133            Delftia      Female           4 0.000000e+00
#> 134       Enterobacter      Female           4 6.485395e-04
#> 135        Escherichia      Female           4 3.270003e-04
#> 136        Haemophilus      Female           4 4.928350e-05
#> 137            Kocuria      Female           4 0.000000e+00
#> 138           Listeria      Female           4 9.230997e-02
#> 139          Moraxella      Female           4 4.901624e-03
#> 140         Mycoplasma      Female           4 2.348193e-03
#> 141 Noviherbaspirillum      Female           4 0.000000e+00
#> 142              Other      Female           4 0.000000e+00
#> 143         Paracoccus      Female           4 1.065317e-01
#> 144        Pseudomonas      Female           4 4.782627e-03
#> 145     Staphylococcus      Female           4 7.375929e-02
#> 146      Streptococcus      Female           4 3.794768e-03
#> 147          Weissella      Female           4 6.277792e-03
#> 148      Acinetobacter        Male           4 5.518243e-01
#> 149          Aeromonas        Male           4 8.681639e-03
#> 150   Alkalihalophilus        Male           4 0.000000e+00
#> 151           Brucella        Male           4 1.538370e-02
#> 152    Corynebacterium        Male           4 8.558884e-02
#> 153      Cutibacterium        Male           4 1.387925e-04
#> 154            Delftia        Male           4 0.000000e+00
#> 155       Enterobacter        Male           4 3.855347e-04
#> 156        Escherichia        Male           4 5.078720e-04
#> 157        Haemophilus        Male           4 1.523616e-03
#> 158            Kocuria        Male           4 0.000000e+00
#> 159           Listeria        Male           4 8.320981e-02
#> 160          Moraxella        Male           4 1.400762e-03
#> 161         Mycoplasma        Male           4 9.850046e-03
#> 162 Noviherbaspirillum        Male           4 0.000000e+00
#> 163              Other        Male           4 0.000000e+00
#> 164         Paracoccus        Male           4 8.275643e-02
#> 165        Pseudomonas        Male           4 9.242495e-04
#> 166     Staphylococcus        Male           4 1.387309e-01
#> 167      Streptococcus        Male           4 3.964288e-03
#> 168          Weissella        Male           4 1.512924e-02
#> 169      Acinetobacter      Female           5 4.208038e-01
#> 170          Aeromonas      Female           5 5.855805e-03
#> 171   Alkalihalophilus      Female           5 0.000000e+00
#> 172           Brucella      Female           5 2.853741e-02
#> 173    Corynebacterium      Female           5 9.247282e-02
#> 174      Cutibacterium      Female           5 1.029391e-03
#> 175            Delftia      Female           5 1.217137e-04
#> 176       Enterobacter      Female           5 1.254729e-03
#> 177        Escherichia      Female           5 2.573825e-03
#> 178        Haemophilus      Female           5 6.746807e-04
#> 179            Kocuria      Female           5 1.887340e-03
#> 180           Listeria      Female           5 1.179112e-01
#> 181          Moraxella      Female           5 9.307237e-04
#> 182         Mycoplasma      Female           5 4.545103e-02
#> 183 Noviherbaspirillum      Female           5 0.000000e+00
#> 184              Other      Female           5 0.000000e+00
#> 185         Paracoccus      Female           5 9.647791e-02
#> 186        Pseudomonas      Female           5 1.919608e-02
#> 187     Staphylococcus      Female           5 1.452855e-01
#> 188      Streptococcus      Female           5 1.546484e-02
#> 189          Weissella      Female           5 4.071162e-03
#> 190      Acinetobacter        Male           5 7.451893e-01
#> 191          Aeromonas        Male           5 0.000000e+00
#> 192   Alkalihalophilus        Male           5 0.000000e+00
#> 193           Brucella        Male           5 1.470818e-02
#> 194    Corynebacterium        Male           5 3.768422e-02
#> 195      Cutibacterium        Male           5 8.509604e-04
#> 196            Delftia        Male           5 0.000000e+00
#> 197       Enterobacter        Male           5 0.000000e+00
#> 198        Escherichia        Male           5 7.292882e-04
#> 199        Haemophilus        Male           5 4.862631e-04
#> 200            Kocuria        Male           5 0.000000e+00
#> 201           Listeria        Male           5 2.436014e-02
#> 202          Moraxella        Male           5 1.896149e-03
#> 203         Mycoplasma        Male           5 1.918182e-02
#> 204 Noviherbaspirillum        Male           5 3.646973e-04
#> 205              Other        Male           5 0.000000e+00
#> 206         Paracoccus        Male           5 3.731782e-02
#> 207        Pseudomonas        Male           5 1.614332e-02
#> 208     Staphylococcus        Male           5 9.094953e-02
#> 209      Streptococcus        Male           5 3.160710e-03
#> 210          Weissella        Male           5 6.977620e-03