Compute linear mixed-effects models (LMM) on longitudinal microbiome data

This function takes an animalcules-formatted MultiAssayExperiment and runs an independent LMM model for each taxon. The model predicts taxon log CPM abundance as a product of fixed-effects covariates with a random effect, usually the unit on which repeated measurements were taken.

Usage

run_lmm_model(
  dat,
  taxon_level = "genus",
  unit_var,
  fixed_cov,
  p_adj_method = "fdr",
  plot_out = FALSE,
  plotsave_loc = ".",
  plot_terms = NULL,
  ...
)

Arguments

dat

A MultiAssayExperiment object specially formatted as an animalcules output.

taxon_level

Character string, default is "genus".

unit_var

Character string giving the variable containing the identifiers for the unit on which multiple measurements were conducted, e.g. subjects. Default is NULL; must be supplied if paired = FALSE.

fixed_cov

A character vector naming covariates to be tested.

p_adj_method

A character string specifying the correction method. Can be abbreviated. See details. Default is "fdr".

plot_out

Logical indicating whether plots should be output alongside the model results. Default is FALSE.

plotsave_loc

A character string giving the folder path to save plot outputs. This defaults to the current working directory.

plot_terms

Character vector. Which terms should be examined in the plot output? Can overlap with the fixed_cov inputs.

...

Further arguments passed to ggsave for plot creation.

Value

A data.frame of modeling results.

Details

P-values are adjusted for the model coefficients within each taxon. The following methods are permitted: c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")

Examples

dat <- system.file("extdata/MAE.RDS", package = "LegATo") |>
  readRDS() |>
  filter_animalcules_MAE(0.05)
out <- run_lmm_model(dat, taxon_level = "genus", unit_var = "Subject",
                     fixed_cov = c("HIVStatus", "timepoint"))
#> [1] "Moraxella"
#> [1] "Dolosigranulum"
#> [1] "Other"
#> [1] "Streptococcus"
#> [1] "Staphylococcus"
#> [1] "Corynebacterium"
#> boundary (singular) fit: see help('isSingular')
#> [1] "Haemophilus"
head(out)
#>   effect Coefficient Coefficient Estimate Standard Error Statistic        df
#> 1  fixed (Intercept)            12857.906       4306.841  2.985461  60.90011
#> 2  fixed (Intercept)            44054.735      26216.459  1.680423  97.97302
#> 3  fixed (Intercept)            19637.811       4845.455  4.052831  42.57545
#> 4  fixed (Intercept)             8408.628       4537.426  1.853171  42.84585
#> 5  fixed (Intercept)            39967.797      11932.245  3.349562  88.22226
#> 6  fixed (Intercept)            31623.496      10208.738  3.097689 164.00000
#>   Unadj p-value           Taxon Adj p-value
#> 1  0.0040731798       Moraxella 0.007128065
#> 2  0.0960609152  Dolosigranulum 0.112071068
#> 3  0.0002106573           Other 0.001474601
#> 4  0.0707525872   Streptococcus 0.099053622
#> 5  0.0011922556  Staphylococcus 0.004172895
#> 6  0.0022950524 Corynebacterium 0.005355122